Split out ListArrayReader into separate module (#1483)
